Fixed 'Are you sure' in delete.pl not to give and error when checked.

+++ cgi-bin/delete.pl
@@ -29,7 +29,7 @@ my $settings_dbh = DBI->connect( "dbi:SQLite:dbname=db/settings", "", "", { Rais
 
 print_page_top_fancy('Delete');
 
-if ( $areyousure eq 'yes' ) {
+if ( defined $areyousure && $areyousure eq 'yes' ) {
    my $drop_table = $settings_dbh->prepare("DELETE FROM settings WHERE name = ?");
    $drop_table->execute($table);
    $drop_table = $settings_dbh->prepare("DELETE FROM modifications WHERE run_id = ?");
